refactor(Calculator): int형의 최대값 연산이 들어와도 합을 구할수 있도록 Calculator의 addition()을 long 타입으로 변경

+++ src/main/java/calculator/controller/CalculatorController.java
@@ -21,11 +21,11 @@ public class CalculatorController {
     public void startCalculate() {
         showStartMessage();
         String input = receiveInput();
-        int result = calculate(input);
+        long result = calculate(input);
         showResultMessage(result);
     }
 
-    private int calculate(String input) {
+    private long calculate(String input) {
         return Calculator.operation(input).addition();
     }
 
@@ -37,7 +37,7 @@ public class CalculatorController {
         outputView.printStartMessage();
     }
 
-    private void showResultMessage(int result) {
+    private void showResultMessage(long result) {
         outputView.printResultMessage(result);
     }
